add new func
add readlist,bookshelf,ratebook,top100book func

+++ src/main/java/com/giang/kesach/resources/IBookShelf.java
@@ -8,6 +8,6 @@ import java.util.List;
 public interface IBookShelf {
     //tra ve sach co trong 1 ke sach
     List<ReadBook> getBookshelfsBook(int id );
-    int addBookToBookShelf(int bookSelfId,int bId);
-    int removeBookFromBookShelf(int bookSelfId,int bId);
+    int addBookToBookShelf(int bookSelfId,long bId);
+    int removeBookFromBookShelf(int bookSelfId,long bId);
 }
